+++ base/exports.jl
@@ -879,6 +879,7 @@ export
     pwd,
     readlink,
     rm,
+    rmrf,
     stat,
     symlink,
     tempdir,